# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Palplus UAA",
        "description": "Palplus Swagger API description",
        "contact": {
            "email": "michaelkinare@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/profile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update Profile",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "description": "Date of Birth",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "physical_address",
                        "in": "query",
                        "description": "physical_address",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/profile/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Retrieve Profile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "activity id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "My Profile",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "description": "Auth Bearer Token Format as 'Bearer <access_token>'",
            "name": "Authorization",
            "in": "header"
        }
    }
}